Set controlling process of all sockets gth.erl exposes to the caller
gth.erl sometimes makes a socket and returns it to the caller, for instance
in new_mtp2_monitor(), new_lapd_layer(), new_player() and new_recorder().

Previously, the API processes retained ownership of the returned socket, thus
preventing the caller from using {active, true} or {active, once}.

Now, the controlling_process is set to the caller, so that {active, once} works
as expected. Passively reading from the socket still works too, of course.