Remove corelatus-internal commands from the sample-code makefile
For a while, Corelatus internal code was kept in CVS and sample
code in Git, with semi-automatic copying going on between the two.
We no longer do that, everything is Git now, and the sample
code is a submodule, so we can ditch the scripts to keep the
two in sync.