Add support for 'verbose' queries in the Erlang sample code
Verbose queries for jobs are supported from gth2_system_36a onwards.
They provide full information about starting arguments for a job, which
is mainly useful for development/debug work.

The new interface is query_jobs/3 and query_job/3.

A side effect of this work is that gth_xml_scan.erl now
supports parsing query replies with multiple items in it,
for example:

   <state><resource name=.../><job .../></state>