C: fix problems found by experimenting with 'clang' -Weverything
- Avoid casting away const
- Make functions and global variables static by default
- Remove unused #defines
- Fix two cases where unintialised variables could be a bug

We don't enable -Weverything by default because it generates
thousands of warnings we're not interested in. If you prefer:

"Nobody should be using -Weverything in their default build flags!
It should only be used as a way to find the names of interesting
warning flags. I’d note that -Weverything even has warnings that
conflict with each other… "
--- Clang developer James Knight